Use the RedirectThread instead
